Data commits

+++ src/app/modules/Type/fee-typeview/fee-typeview.component.ts
@@ -1,7 +1,9 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatTableDataSource } from '@angular/material/table';
 import { Router } from '@angular/router';
+import { ConfirmationDialogComponent } from 'src/app/shared/confirmation-dialog/confirmation-dialog.component';
 import { UtilityServiceService } from 'src/app/utility-service.service';
 
 @Component({
@@ -18,11 +20,11 @@ export class FeeTypeviewComponent implements OnInit {
 
     constructor(
         private _dataService: UtilityServiceService,
-        private router: Router
+        private router: Router,
+        public dialog: MatDialog
     ) { }
 
     ngOnInit() {
-
         this._dataService.getFeetypeList().subscribe(res => {
             const result = res.filter(e => !!e.id);
             this.dataSource = new MatTableDataSource(result);
@@ -31,9 +33,19 @@ export class FeeTypeviewComponent implements OnInit {
     }
 
     deleteSchool(id) {
-        this._dataService.deleteFeeType(id).subscribe(res => {
-            this.ngOnInit();
+        const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+            width: '300px',
+            data: {}
         });
+
+        dialogRef.afterClosed().subscribe(result => {
+            if (result) {
+                this._dataService.deleteFeeType(id).subscribe(res => {
+                    this.ngOnInit();
+                });
+            }
+        });
+
     }
 
     updateFeeType(element) {
